close writers from upstream
If a response would've been answered in part by upstream
and in part from cache the pipeWriter for the upstream writer would not be closed
as it's a NopWriterCloser. This leads to the pipeReader not getting EOF
and the MultiReadCloser waiting for more reads when there would be no
more.

Implementing io.Closer for FlexibleResponseWriter with the Close method
that calls the one on the enclosed BodyWriter as well as using a new
ResponseWriteCloser interface ensures that whereever possible a non noop
Close method will be called.