add RemoveObject and RemoveObjectsForKey
both functions remove multiple object indexes from the algorithm and
call the underlying remove provided remove function.

In the case of the cache.lru implementation both functions will lock the
whole Algorithm instance and depending on the underlying implementation
of the remove function this might be for a long time.