Add types App and AppStats
App is an interface with a single method. The interfaceness comes from
the need to not make import cycles. The interface is meant to be extented
when(if) it's required in the future.

AppStats idea is to be used to collect the number of  requests
against the application. As well as both the responses that have been
handled, as well as those that haven't been (returning 404) due to
no combination of virtualhosts, aliases and ocations corresponded to them.

The reason AppStats is not directly added to a context before giving it
to the next handler is that it's rarely needed. Adding the Application
(implementing the App) to the application wide context will make it so
that only in the places where it's required memory for AppStats will be
allocated as well as less contexts will be created.