absolutize os.Args[0]
grace uses os.Args[0] to find out what executable to run upon restart.
It uses exec.LookPath which is fine, but as nedomi changes it's working
dir a relative path may/will not point to the same file as it was
intented. Making the path absolute fixes all this.

If that is a relative path and the working dir has been changed - this
will not work (or atleast not as wanted). Making the path absolute fixes the issue.