fix panic in GetAvailableParts when no files are found
If no files were found a slice with capacity '-1' would be tried to be
made - which will fail with panic. The -1 was in order for the metadata
file to not be counted toward the parts size - saving the memory for a
pointer in case Metadata is available, using double the amount of
memory if it's not there - an optimization that is probably pointless
and so removed.