Separates IsResponseCacheable in two different functions
A second function is created, ResponseExpiresIn, in effort to
separate the logic between deciding if something is cacheable and
deciding for how long something can be cached.

IsResponseCacheable now returns a bool wich means "yes, it is
cacheable" and "no, it is not cacheable" for a response.

ResponseExpiresIn uses the response headers to decide for how long
could something be cached. It has a second parameter which will
be used as default duration in case nothing else is found.