cache.handler: return 500 on 'too many open files'
Up until now on 'too many open files' a new connection to the upstream
was opened which probably would fail as well.

This change does not try to make it less likely for 'too many open files'
to occur, but instead cuts short the request handling and returns an
InternalServerError where possible or just stop returning data if
already in the middle of a response.