added headers rewriter
The added handler 'headers' could be used to rewrite both the response's
and the request's headers.

The added config options in the http, vhost and location section will be
used to make a headers handler for the incoming requests before any handler
in chain is called. The settings are "inherited" in the inner settings.