reload storage and algorithm logger as well as more settings
up until now upon reload the loggers for reused storages and algorithms
were not changed/reloaded. This commit fixes that.

This was found while adding two new configurations for Algorithm:
BulkRemoveCount and BulkRemoveTimeout(in milliseconds). Which are also
part of the commit because of time constraints.

Because of the multitude of changes and for future proving the function
to reload the logger was named ChangeConfigs. The lru algorithm
implementation really changes it's config which probably should be done
from the outside and then the call to mean to just reload the config.

That will take more rewriting for the types.Logger reloading and as
such is left for the future.