change http.Server.*Timeout to be on a per Read|Write basis
This is fun feature which circumvents the fact that http.Server timeout
is not useful when you have long living connections. If you have a
connection which will have to live for 20 minutes you can either put
your timeout to something above 20 minutes and possibly have connections
which are stale for 20 minutes, have timeout connections for no good
reason or do what is done through this commit.

This commit changes the meaning of net.Conn.Set*Deadline to mean
"calculate me a timeout from now till the deadline and than use that to
deadline each Read|Write operation for that timeout", not all of them.

This means that you can have a timeout of 10 seconds and connections
which have been used to transmit data for as long as each Read|Write
takes less than 10 seconds.

This is only done for the connections served by facebook's grace - the
incoming traffic for nedomi. For this reason the grace project was ...
forked. Vendoring the gracehttp package should be considered at a future
date.