misc changes to conform with reference ketama implementation
in the reference implementation only the paths are being md5ed and than
hex to stringed and than md5ed. The upstreams are just md5ed and turned
into 32bit int

also sort.Search and the reference implementation search were giving
different result.

Less is not actually less not more ;)