Vhost is identified by Host header in access logs
Previously in the access log line for a request the primary vhost
name was used no matter what. Accessing the vhost with any of the
aliases was hidden. Now the host header is preferred when available.

This has the side effect that all requests for vhosts which are not
configured will potentionally have different Host headers.