Git vars can be overriden wint env vars
If you are building in a gitless directory the git commands will
fail. This may be because you are building from a source archive.

This commit gives you the option to supply the variables
manually via environment variables:

    $ GIT_TAG="v0.1.2" make

This does not solve the problem completely since ideally,
you should be able to just type `make` and compile nedomi
without warnings.