fix a race condition around the app's logger
previously when the logger was changed, due to reload, every other part
using the logger was being told seperately. This includes the
CacheAlgorithm, Storage and (since a few commits) the Scheduler.

Although the Application was synchronizing around the it's logger,
all other components never did. Now every component synchronizes with
the new new types.SyncLogger (naming is hard).