handler.dir: gets `strip` setting
this is mostly useful(and semi required) when the dir handler is used in
a location with a path different from `/` for example `/doc`.
Without it a request `/var/www/index.html` (with dir handler with `root`
`/var/www`) will try to serve file with path `/var/www/doc/index.html`.

To fix this the new `strip` setting was added. It will strip a provided
path from the beginning of the requested path.

Setting it to `/doc` in the above example will give the (expected)
result of serving `/var/www/index.html`.